requirements.txt
Django==4.0.2
djangorestframework==3.14.0
